Refactor redirectcomponent generation for docs
