+++ lib/server/docs.js
@@ -93,8 +93,28 @@ function getComponent(rawContent, mdToHtml, metadata) {
   );
 }
 
+function getRedirectComponent(metadata) {
+  if (
+    !env.translation.enabled ||
+    metadata.permalink.indexOf('docs/en') === -1
+  ) {
+    return null;
+  }
+  const Redirect = require('../core/Redirect.js');
+  const redirectlink = getPath(metadata.permalink, siteConfig.cleanUrl);
+  return (
+    <Redirect
+      metadata={metadata}
+      language={metadata.language}
+      config={siteConfig}
+      redirect={siteConfig.baseUrl + redirectlink}
+    />
+  );
+}
+
 module.exports = {
   getComponent,
   getFile,
+  getRedirectComponent,
   mdToHtmlify,
 };
