+++ lib/server/generate.js
@@ -7,16 +7,13 @@
 
 async function execute() {
   require('../write-translations.js');
-
   const metadataUtils = require('./metadataUtils');
   const blog = require('./blog');
   const docs = require('./docs');
-
   const CWD = process.cwd();
   const fs = require('fs-extra');
   const readMetadata = require('./readMetadata.js');
   const path = require('path');
-  const {getPath} = require('../core/utils.js');
   const {minifyCss, isSeparateCss} = require('./utils');
   const React = require('react');
   const mkdirp = require('mkdirp');
@@ -73,10 +70,7 @@ async function execute() {
   // needed when the project's a GitHub org page
 
   const buildDir = join(CWD, 'build', siteConfig.projectName);
-
   const mdToHtml = metadataUtils.mdToHtml(Metadata, siteConfig.baseUrl);
-
-  const Redirect = require('../core/Redirect.js');
 
   fs.removeSync(join(CWD, 'build'));
 
@@ -94,28 +88,16 @@ async function execute() {
     writeFileAndCreateFolder(targetFile, str);
 
     // generate english page redirects when languages are enabled
-    if (
-      env.translation.enabled &&
-      metadata.permalink.indexOf('docs/en') !== -1
-    ) {
-      const redirectlink = getPath(metadata.permalink, siteConfig.cleanUrl);
-      const redirectComp = (
-        <Redirect
-          metadata={metadata}
-          language={metadata.language}
-          config={siteConfig}
-          redirect={siteConfig.baseUrl + redirectlink}
-        />
-      );
-      const redirectStr = renderToStaticMarkupWithDoctype(redirectComp);
-
-      // create a redirects page for doc files
-      const redirectFile = join(
-        buildDir,
-        metadata.permalink.replace('docs/en', 'docs')
-      );
-      writeFileAndCreateFolder(redirectFile, redirectStr);
+    const redirectComp = docs.getRedirectComponent(metadata);
+    if (!redirectComp) {
+      return;
     }
+    const redirectStr = renderToStaticMarkupWithDoctype(redirectComp);
+    const redirectFile = join(
+      buildDir,
+      metadata.permalink.replace('docs/en', 'docs')
+    );
+    writeFileAndCreateFolder(redirectFile, redirectStr);
   });
 
   // copy docs assets if they exist
@@ -145,7 +127,6 @@ async function execute() {
       if (extension !== '.md' && extension !== '.markdown') {
         return;
       }
-
       const urlPath = blog.fileToUrl(normalizedFile);
       const str = blog.getPost(normalizedFile, urlPath, siteConfig);
       const targetFile = join(buildDir, 'blog', urlPath);
